Merge tag 'tags/ArduPlane-3.2.2' into Soar_lib

Conflicts:
	ArduPlane/ArduPlane.pde
	ArduPlane/Log.pde
	ArduPlane/defines.h
	README.md

+++ libraries/AP_GPS/examples/GPS_AUTO_test/GPS_AUTO_test.pde
@@ -34,6 +34,8 @@
 #include <AP_Notify.h>
 #include <AP_BoardLED.h>
 #include <AP_NavEKF.h>
+#include <AP_Rally.h>
+#include <AP_Scheduler.h>
 
 const AP_HAL::HAL& hal = AP_HAL_BOARD_DRIVER;
 
@@ -70,7 +72,7 @@ void loop()
                             (int)gps.ground_course_cd() / 100,
                             gps.num_sats(),
                             gps.time_week(),
-                            gps.time_week_ms(),
+                            (unsigned long)gps.time_week_ms(),
                             gps.status());
     }
     hal.scheduler->delay(10);
